+++ CHANGELOG.md
@@ -11,6 +11,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Conviva
 	- Added average bitrate reporting
+### Fixed
+
+- Conviva
+  - Improved the VPF detection for iOS and tvOS.
+
+### Removed
+
+- Conviva
+  - Removed Objective-C API support.
+
+- Yospace
+  - Removed Objective-C API support.
 
 ## [8.13.0] - 2025-03-13